requirements.txt
dash>=2.17.0
dash-bootstrap-components>=1.6.0
plotly>=5.22.0
anthropic>=0.28.0
langgraph>=0.2.0
langchain-anthropic>=0.1.15
langchain-core>=0.2.0
python-pptx>=0.6.23
openpyxl>=3.1.2
pandas>=2.2.0
